requirements.txt
torch>=1.12
torchvision>=0.13
numpy>=1.21
Pillow
matplotlib
tqdm
tensorboard 